common/remoteconnection.h,net/progclient.cc,net/remoteconnection.cc: GetLastError() has an unsigned return type, so need to cast to int before negating.
git-svn-id: file:///home/xapian-svn/xapian/trunk@8226 6b2f1b92-a4f3-0310-943f-de3cbe41748b

+++ xapian-core/common/remoteconnection.h
@@ -69,7 +69,7 @@ struct WinsockInitializer {
  *  the return value will be the value of errno.
  */
 inline int socket_errno() {
-    return -WSAGetLastError();
+    return -(int)WSAGetLastError();
 }
 #else
 // Use a macro so we don't need to pull safeerrno.h in here.
